Now the bot is running and working properly, now onto the first polishing pass

+++ opzbackup/OPZBackup/FileManagement/FileCleaner.cs
@@ -2,8 +2,13 @@
 
 public static class FileCleaner
 {
-    public static async Task DeleteDirAsync(string channelDirPath)
+    public static async Task<bool> DeleteDirAsync(string channelDirPath)
     {
+        //TODO-3 Should have a logger here too
+        if (!Directory.Exists(channelDirPath))
+            return false;
+
         await Task.Run(() => Directory.Delete(channelDirPath, recursive: true));
+        return true;
     }
 }
